+++ RELEASE_NOTES.md
@@ -6,6 +6,7 @@
 - fix: Populate TestResult.ResourceCount in `/verificationLogger:csv` output correctly when verification condition splitting occurs (e.g. when using `/vcsSplitOnEveryAssert`).
 - fix: DafnyOptions.Compiler was null, preventing instantiation of ModuleExportDecl (https://github.com/dafny-lang/dafny/pull/1933)
 - fix: /showSnippets crashes Dafny's legacy server (https://github.com/dafny-lang/dafny/pull/1970)
+- break: Change the namespace of the C# Dafny runtime from `Dafny` to `DafnyRuntime` (https://github.com/dafny-lang/dafny/pull/1999)
 
 # 3.5.0